Move page number functions into its own file
